refactor: group the components related to NewGroup
