+++ app/page.js
@@ -4,10 +4,7 @@ import MapWrapper from "@/components/MapWrapper";
 import Groups from "@/components/Groups";
 import SearchBar from "@/components/SearchBar";
 import SearchBarById from "@/components/SearchBarById";
-import NewGroupButton from "@/components/NewGroupButton";
-import NewGroupTime from "@/components/NewGroupTime";
-import NewGroupDate from "@/components/NewGroupDate";
-import NewGroupLimit from "@/components/NewGroupLimit";
+import NewGroup from "@/components/NewGroup/NewGroup";
 
 export default function Home() {
   // isNewGroup? -> render other components
@@ -16,10 +13,7 @@ export default function Home() {
       <MapWrapper />
       <SearchBarById />
       <Groups />
-      <NewGroupButton />
-      <NewGroupTime />
-      <NewGroupDate />
-      <NewGroupLimit />
+      <NewGroup />
     </div>
   );
 }
