+++ components/NewGroup/Time.jsx
@@ -3,7 +3,7 @@ import InputLabel from "@mui/material/InputLabel";
 import MenuItem from "@mui/material/MenuItem";
 import FormControl from "@mui/material/FormControl";
 import Select from "@mui/material/Select";
-import styles from "../styles/newGroupTime.module.scss";
+import styles from "../../styles/newGroupTime.module.scss";
 
 export default function NewGroupTime() {
   const hourOptions = [
